// src/types/r3eTypes.ts
export enum ESession {
	Unavailable = -1,
	Practice = 0,
	Qualify = 1,
	Race = 2,
	Warmup = 3
}
// ...
export interface ICutTrackPenalties {
	DriveThrough: number;
	StopAndGo: number;
	PitStop: number;
	TimeDeduction: number;
	SlowDown: number;
}